Reduced character flashing
Increased the character movement rates and added a delay. The issue was
that the functions were drawing a new image for each pixel the character
had to move (this is also why he was so slow) and this put a lot of load
on the processor. I decreased this load by a factor of four. More
optimization code to come. It may be that the icon redisplay system is
not fast enough to run the game smoothly. In any case, that will have to
be reworked, as it does not work on built jarfiles.