Major Bugfixes
Fixed an issue where objects were staying in the
world after they had been destroyed. The issue
was due to the fact that they were not removed
from the Tile.objects array in the destructor.
There was also a reference error in GameObject
which prevented imageIDs from being set properly.